chore: move preflight checks into a separate package
