+++ internal/controller/consts.go
@@ -16,13 +16,7 @@ limitations under the License.
 
 package controller
 
-import "time"
-
 const (
-	// preflightFailedRequeueAfter is how long to wait before trying to reconcile
-	// if some preflight check has failed.
-	preflightFailedRequeueAfter = 30 * time.Second
-
 	httpsScheme             = "https"
 	githubDomain            = "github.com"
 	gitlabHostPrefix        = "gitlab."
